feat/74/club-update

+++ 2nd-backend/src/main/java/com/ohgiraffers/secondbackend/readingclub/repository/ReadingClubRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ReadingClubRepository extends JpaRepository<ReadingClub, Integer> {
+public interface ReadingClubRepository extends JpaRepository<ReadingClub, Long> {
 
 }
